Make the strict parsing test fixture import safe.
Previously, the fixture to test module parsing referred to an unresolvable
import. This now causes the parse to fail by default. Update the code to
trigger module behaviour without failing in the resolver. The key property is
the presence of import/export statements.

At the same time, suppress a complaint from the more recent versions of AVA,
that there were no tests in the fixture, by adding a trivial test that always
passes. Since the purpose of this fixture is to parse successfully, the ability
to execute the test at all means it passed.

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>